+++ storage/visitorVariation.go
@@ -3,14 +3,14 @@ package storage
 import "time"
 
 type VisitorVariation struct {
-	VariationId    int
-	AssignmentDate int64
+	VariationId    uint32
+	AssignmentDate uint32
 }
 
-func NewVisitorVariation(variationId int) *VisitorVariation {
-	return &VisitorVariation{VariationId: variationId, AssignmentDate: time.Now().Unix()}
+func NewVisitorVariation(variationId uint32) *VisitorVariation {
+	return &VisitorVariation{VariationId: variationId, AssignmentDate: uint32(time.Now().Unix())}
 }
 
-func (variation VisitorVariation) isValid(respoolTime int64) bool {
+func (variation VisitorVariation) isValid(respoolTime uint32) bool {
 	return variation.AssignmentDate > respoolTime
 }