+++ utils/array_utils.go
@@ -9,21 +9,31 @@ func ArrayToString(a []int, delim string) string {
 	return strings.Trim(strings.Replace(fmt.Sprint(a), " ", delim, -1), "[]")
 }
 
-func Contains(s []string, str string) bool {
+func Contains(s []interface{}, str string) bool {
 	for _, v := range s {
-		if v == str {
-			return true
+		if _, ok := v.(string); ok {
+			if v == str {
+				return true
+			}
 		}
 	}
 	return false
 }
 
-func Map(vs []string, f func(string) string) []string {
-	vsm := make([]string, len(vs))
-	for i, v := range vs {
-		vsm[i] = f(v)
+func Map(array []interface{}, f func(interface{}) interface{}) []interface{} {
+	mapArray := make([]interface{}, len(array))
+	for i, v := range array {
+		mapArray[i] = f(v)
 	}
-	return vsm
+	return mapArray
+}
+
+func MapToStringDict(array []interface{}, f func(*map[string]interface{}, interface{})) map[string]interface{} {
+	mapDict := make(map[string]interface{}, len(array))
+	for _, v := range array {
+		f(&mapDict, v)
+	}
+	return mapDict
 }
 
 func Remove(sequence interface{}, index int) interface{} {